pyproject.toml
[project]
name = "questions-api"
version = "0.1.0"
description = "A simple REST API for solving questions"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.4",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-spectacular>=0.28.0",
    "psycopg[binary]>=3.2.9",
    "python-dotenv>=1.1.1",
]
